pyproject.toml
[tool.poetry]
name = "attrs-serde"
version = "0.1.1"
authors = ["Dotan Nahum <jondotan@gmail.com>"]
description = "A serialization addon for attrs"

license = "MIT"

readme = "README.md"

homepage = "https://github.com/jondot/attrs-serde"
repository = "https://github.com/jondot/attrs-serde"
documentation = "https://github.com/jondot/attrs-serde"

keywords = ["attrs", "serialization", "dict", "data", "nested", "functional"]

[tool.poetry.dependencies]
python = "*"
toolz = "^0.9.0"
cytoolz = "^0.9.0"
attrs = "^18.2"

[tool.poetry.dev-dependencies]
boltons = "^18.0"
pytest = "^3.0"
pytest-benchmark = "^3.1"
pytest-testmon = "^0.9.13"
pytest-watch = "^4.2"
snapshottest = "^0.5.0"

flake8 = { version = "^3.5", optional = true }
flake8-mypy = { version = "^17.8", optional = true }
pytest-jest = "^0.3.0"

[tool.poetry.extras]
lint = ["flake8", "flake8-mypy"]